tests/tools: refactor common tools TestCall unit tests
