+++ gateway_code/utils/tests/avrdude_test.py
@@ -28,7 +28,6 @@
 # pylint: disable=maybe-no-member
 # pylint: disable=unused-argument
 
-import time
 import os.path
 import unittest
 
@@ -59,37 +58,6 @@ class TestsMethods(unittest.TestCase):
     def test_invalid_firmware_path(self):
         ret = self.avr.flash('/invalid/path')
         self.assertNotEqual(0, ret)
-
-
-class TestsCall(unittest.TestCase):
-    """ Tests avrdude call timeout """
-    def setUp(self):
-        self.timeout = 5
-        self.avr = avrdude.AvrDude(NodeLeonardo.AVRDUDE_CONF,
-                                   timeout=self.timeout)
-        self.avr.args = mock.Mock()
-
-    def test_timeout_call(self):
-        """Test timeout reached."""
-        self.avr.args.return_value = {'args': ['sleep', '10']}
-        t_0 = time.time()
-        ret = self.avr.call_cmd('sleep')
-        t_end = time.time()
-
-        # Not to much more
-        self.assertLess(t_end - t_0, self.timeout + 1)
-        self.assertNotEqual(ret, 0)
-
-    def test_no_timeout(self):
-        """Test timeout not reached."""
-        self.avr.args.return_value = {'args': ['sleep', '1']}
-        t_0 = time.time()
-        ret = self.avr.call_cmd('sleep')
-        t_end = time.time()
-
-        # Strictly lower here
-        self.assertLess(t_end - t_0, self.timeout - 1)
-        self.assertEqual(ret, 0)
 
 
 @mock.patch('serial.Serial')
