+++ gateway_code/utils/tests/cc2538_test.py
@@ -21,7 +21,6 @@
 """" some tests for the CC2538 wrapper"""
 
 import os
-import time
 import unittest
 
 import mock
@@ -78,36 +77,3 @@ class TestsCC2538Methods(unittest.TestCase):
         """Test flash an invalid firmware return a non zero value."""
         ret = self.cc2538.flash('/invalid/path')
         assert ret > 0
-
-
-class TestsCC2538Call(unittest.TestCase):
-    # pylint:disable=protected-access
-    """ Tests cc2538-bsl call timeout """
-    def setUp(self):
-        self.timeout = 5
-        self.cc2538 = cc2538.CC2538({'port': NodeFirefly.TTY,
-                                     'baudrate': NodeFirefly.BAUDRATE},
-                                    timeout=self.timeout)
-        self.cc2538.args = mock.Mock()
-
-    def test_timeout_call(self):
-        """Test timeout reached."""
-        self.cc2538.args.return_value = {'args': ['sleep', '10']}
-        t_0 = time.time()
-        ret = self.cc2538.call_cmd('sleep')
-        t_end = time.time()
-
-        # Not to much more
-        self.assertLess(t_end - t_0, self.timeout + 1)
-        self.assertNotEqual(ret, 0)
-
-    def test_no_timeout(self):
-        """Test timeout not reached."""
-        self.cc2538.args.return_value = {'args': ['sleep', '1']}
-        t_0 = time.time()
-        ret = self.cc2538.call_cmd('sleep')
-        t_end = time.time()
-
-        # Strictly lower here
-        self.assertLess(t_end - t_0, self.timeout - 1)
-        self.assertEqual(ret, 0)
